# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Patoughi Driver API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/transports/{transport}": {
            "get": {
                "tags": [
                    "Transport"
                ],
                "summary": "Get transport details",
                "operationId": "getTransportDetail",
                "parameters": [
                    {
                        "name": "transport",
                        "in": "path",
                        "description": "Transport ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TransportDetailResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/transports": {
            "get": {
                "tags": [
                    "Transport"
                ],
                "summary": "List all transports assigned to driver",
                "description": "Retrieves a paginated list of transports filtered by status (e.g., \"assigned\").",
                "operationId": "listTransports",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved transports",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TransportResource"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication failed (invalid/missing token)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorUnauthorized"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error (e.g., invalid status or pagination parameters)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/vehicles/{vehicle}/change-status": {
            "patch": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Update vehicle Change status",
                "operationId": "updateVehicleChangeStatus",
                "parameters": [
                    {
                        "name": "vehicle",
                        "in": "path",
                        "description": "Vehicle ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update vehicle Change Status",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehicleStatusUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Vehicle Updated Successfully"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "success": {
                                            "type": "string",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorValidation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseErrorServer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "VehicleStatusUpdateRequest": {
                "required": [],
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "maintenance",
                            "retired",
                            "inactive"
                        ]
                    }
                },
                "type": "object"
            },
...
            "BidResource": {
                "required": [
                    "id",
                    "driver_id",
                    "transport_id",
                    "proposed_price",
                    "description",
                    "status"
                ],
                "properties": {
                    "driver_id": {
                        "type": "string"
                    },
                    "transport_id": {
                        "type": "string"
                    },
                    "proposed_price": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "accepted",
                            "rejected",
                            "pending",
                            "cancelled"
                        ]
                    }
                },
                "type": "object"
            },
            "CargoTypeResource": {
                "required": [
                    "id",
                    "name",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "DispatchUnitTypeResource": {
                "required": [
                    "id",
                    "name",
                    "short_name"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "short_name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "OrderResource": {
                "required": [
                    "id",
                    "reference_number",
                    "status",
                    "description",
                    "total_price_value",
                    "total_price_currency",
                    "origin_loading_location",
                    "destination_loading_location",
                    "expected_delivery_date",
                    "allow_price_negotiation"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "reference_number": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "initiated",
                            "confirmed",
                            "processing",
                            "delivered",
                            "canceled"
                        ]
                    },
                    "description": {
                        "type": "string"
                    },
                    "total_price_value": {
                        "type": "decimal"
                    },
                    "total_price_currency": {
                        "type": "string"
                    },
                    "origin_loading_location": {
                        "$ref": "#/components/schemas/LoadingLocationResource"
                    },
                    "destination_loading_location": {
                        "$ref": "#/components/schemas/LoadingLocationResource"
                    },
                    "expected_delivery_date": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "allow_price_negotiation": {
                        "type": "boolean"
                    },
                    "shipments_count": {
                        "type": "string"
                    },
                    "transports": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TransportResource"
                        }
                    },
                    "shipments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ShipmentResource"
                        }
                    }
                },
                "type": "object"
            },
            "ShipmentResource": {
                "required": [
                    "id",
                    "tracking_number",
                    "order_id",
                    "status",
                    "name",
                    "dispatch_unit_type_id",
                    "expected_pickup_at",
                    "expected_drop_off_at",
                    "total_weight",
                    "total_volume",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "tracking_number": {
                        "type": "string"
                    },
                    "order_id": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "initiated",
                            "confirmed",
                            "processing",
                            "delivered",
                            "canceled"
                        ]
                    },
                    "name": {
                        "type": "string"
                    },
                    "dispatch_unit_type": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/DispatchUnitTypeResource"
                            }
                        ],
                        "nullable": true
                    },
                    "cargo_type": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/CargoTypeResource"
                            }
                        ],
                        "nullable": true
                    },
                    "expected_pickup_at": {
                        "format": "date-time"
                    },
                    "expected_drop_off_at": {
                        "type": "date-time"
                    },
                    "total_weight": {
                        "type": "number",
                        "format": "double"
                    },
                    "total_volume": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "TransportDetailResource": {
                "required": [
                    "id",
                    "status",
                    "order",
                    "shipments",
                    "bids"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "example": "AWAITING_BID"
                    },
                    "order": {
                        "$ref": "#/components/schemas/OrderResource"
                    },
                    "shipments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ShipmentResource"
                        }
                    },
                    "bids": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BidResource"
                        }
                    }
                },
                "type": "object"
            },
            "TransportResource": {
                "required": [
                    "id",
                    "status",
                    "vehicle_type",
                    "vehicle",
                    "shipments",
                    "order",
                    "assignment_status",
                    "waybills"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "01JTZQ30"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "clarify_pickup",
                            "in_transit",
                            "passed_stopover",
                            "delivered",
                            "completed",
                            "exception",
                            "investigate_incomplete_delivery",
                            "closed_cancelled",
                            "closed_with_error",
                            "closed"
                        ]
                    },
                    "vehicle_type": {
                        "$ref": "#/components/schemas/VehicleTypeResource"
                    },
                    "vehicle": {
                        "$ref": "#/components/schemas/VehicleResource"
                    },
                    "order": {
                        "$ref": "#/components/schemas/OrderResource"
                    },
                    "shipments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ShipmentResource"
                        }
                    },
                    "waybills": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/WaybillResource"
                        }
                    },
                    "assignment_status": {
                        "type": "string",
                        "enum": [
                            "assigned",
                            "unassigned",
                            "cancelled"
                        ]
                    }
                },
                "type": "object"
            },
...
            "VehicleTypeResource": {
                "required": [
                    "id",
                    "name",
                    "name_en",
                    "vehicle_length",
                    "vehicle_width",
                    "capacity_pal",
                    "capacity_weight",
                    "max_loading_height",
                    "trip_type",
                    "rtmp_code",
                    "rtmp_min_capacity",
                    "rtmp_max_capacity"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "name_en": {
                        "type": "string"
                    },
                    "vehicle_length": {
                        "type": "number",
                        "format": "double"
                    },
                    "vehicle_width": {
                        "type": "number",
                        "format": "double"
                    },
                    "capacity_pal": {
                        "type": "integer"
                    },
                    "capacity_weight": {
                        "type": "integer"
                    },
                    "max_loading_height": {
                        "type": "integer"
                    },
                    "trip_type": {
                        "type": "string",
                        "enum": [
                            "single",
                            "round",
                            "multi"
                        ]
                    },
                    "rtmp_code": {
                        "type": "integer"
                    },
                    "rtmp_min_capacity": {
                        "type": "integer"
                    },
                    "rtmp_max_capacity": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "WaybillResource": {
                "required": [
                    "id",
                    "driver",
                    "transport",
                    "proposed_price"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "01JTZQ30ZR13VDZGMQCHDCB5MB"
                    },
                    "driver": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": {
                                "id": 1,
                                "full_name": "Abbass Alavy"
                            }
                        }
                    },
                    "transport": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": {
                                "id": 1,
                                "full_name": "Abbass Alavy"
                            }
                        }
                    },
                    "proposed_price": {
                        "type": "string",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "ApiResponseErrorUnauthorized": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Server error message"
                    },
                    "errors": {
                        "description": "Validation errors",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ApiResponseErrorValidation": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Server error message"
                    },
                    "errors": {
                        "description": "Validation errors",
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ApiResponseErrorServer": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Server error message"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}